+++ crates/codexi-cli/src/command/mod.rs
@@ -12,6 +12,7 @@ mod loan;
 mod operation;
 mod report;
 mod root;
+mod settings;
 
 pub use account::AccountCommand;
 pub use admin::{AdminCommand, TrashCommand};
@@ -25,3 +26,4 @@ pub use loan::{LoanCommand, LoanPolicyAction};
 pub use operation::OperationCommand;
 pub use report::ReportCommand;
 pub use root::{Cli, RootCommand};
+pub use settings::SettingsCommand;